move param measures to convert
