+++ specparam/tests/convert/test_periodic.py
@@ -1,9 +1,17 @@
-"""Test functions for specparam.convert.params."""
+"""Test functions for specparam.convert.periodic."""
 
-from specparam.convert.params import *
+from specparam.convert.periodic import *
 
 ###################################################################################################
 ###################################################################################################
+
+def test_compute_fwhm():
+
+    assert compute_fwhm(1.5)
+
+def test_compute_gauss_std():
+
+    assert compute_gauss_std(1.0)
 
 def test_compute_peak_height(tfm):
